git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@18023 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/cache/cachelib.php
@@ -100,7 +100,7 @@ class Cachelib {
 	$back = array();
 	$all = opendir($path);
 	while ($file = readdir($all)) {
-		if (is_dir($path.'/'.$file) and $file <> ".." and $file <> "." and $file <> "CVS") {
+		if (is_dir($path.'/'.$file) and $file <> ".." and $file <> "." and $file <> "CVS" and $file <> ".svn" ) {
 			$du = $this->du($path.'/'.$file);
 			$total+= $du['total'];
 			$cant+= $du['cant'];
@@ -125,7 +125,7 @@ class Cachelib {
 	if (!$path or !is_dir($path)) return 0;
 	if ($dir = opendir($path)) {
 		while (false !== ($file = readdir($dir))) {
-			if (substr($file,0,1) == "." or $file == 'CVS' or $file == "index.php" or $file == "README" ) continue;
+			if (substr($file,0,1) == "." or $file == 'CVS' or $file == '.svn' or $file == "index.php" or $file == "README" ) continue;
 			if (is_dir($path."/".$file)) {
 				$this->erase_dir_content($path."/".$file);
 				rmdir($path."/".$file);
